+++ IHardwareProvider.h
@@ -7,13 +7,13 @@ class IHardwareProvider {
     virtual void initialize() {}
 
     virtual void applyCorrection(Vector3 correction) {}
-    virtual void throttleMotors(float throttlePercentage) {}
+    virtual void throttleMotors(double throttlePercentage) {}
     virtual void deployLegs(bool deploy = true) {}
-    
+
     virtual void updateWiggle(unsigned long timeLeftMillis) {}
     virtual void updateCountdown(unsigned long timeLeftMillis) {}
-    
+
     virtual void lightLed(int r, int g, int b) {}
-    virtual void toneBuzzer(float frequency) {}
+    virtual void toneBuzzer(double frequency) {}
     virtual void stopBuzzer() {}
 };
